+++ src/bot/Options.ts
@@ -1,9 +1,9 @@
 import { Guild } from "discord.js";
-import { DeepPartialExcept, NonEmptyReadonly, Overwrite, PromiseOrSync } from "./utils";
-import { BotDatabaseHandler, Property } from "./database";
-import { JsonDatabaseHandler } from "./builtIn/database";
-import { LoggerOptions } from "./Logger";
-import { pipeLoggerTheme } from "./builtIn/loggerTheme/Pipe";
+import { DeepPartialExcept, NonEmptyReadonly, Overwrite, PromiseOrSync } from "../utils";
+import { BotDatabaseHandler, Property } from "../database";
+import { JsonDatabaseHandler } from "../builtIn/database";
+import { LoggerOptions } from "../Logger";
+import { pipeLoggerTheme } from "../builtIn/loggerTheme/Pipe";
 import { Bot } from "./Bot";
 
 /**
