+++ src/bot/index.ts
@@ -1,0 +1,2 @@
+export * from "./Bot";
+export * from "./Options";
